V0.19: frequency counter improvements
There are disturbances caused by the seven-segment displays. Better power-supply decoupling and separate power-supply for these SSDs improve the signal. The frequency counter has also been made more robust, but will still not always give a proper reading. It now displays a 0 if the actual frequency could not be determined.